Motor Control
Final step of autonomous rc car driving. This package is not completed yet because there is some issue with controlling the speed of motor of rc car. The conventional rc car motor is controlled by limiting the current supplied. However, using motor driver and arduino, we can only control the voltage supplied. Further research needs to be done.